LayoutTests/svg/dynamic-updates/SVGFilterElement-svgdom-height-prop-expected.txt

CONSOLE MESSAGE: line 43: TypeError: 'undefined' is not an object (evaluating 'filterElement.height.baseVal')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'height' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL filterElement.height.baseVal.value should be 100. Threw exception TypeError: 'undefined' is not an object (evaluating 'filterElement.height.baseVal')
